+++ src/main/java/com/acme/repository/CertificationRepository.java
@@ -18,7 +18,7 @@ public interface CertificationRepository {
 	public List<Certification> searchCertificationFromFamilyProfessional(FamilyProfessional f);
 	public List<FamilyProfessional> getAllFamilyProfessional();
 	public FamilyProfessional getFamilyProfessionalByName(String name);
-	public FamilyProfessional getFamilyProfessionalById(Integer id);
+	public FamilyProfessional getFamilyProfessionalById(Long id);
 	public void removeCertification(Certification c);
 	public void removeExam(Exam e);
 	public void removeFProfessional(FamilyProfessional fp);
@@ -29,5 +29,5 @@ public interface CertificationRepository {
 	public void updateFProfessional(FamilyProfessional fp);
 	public void persistPreRegister(PreRegister p);
 	public void updatePreRegister(PreRegister  p);
-	public Certification getCertificationById(Integer id);
+	public Certification getCertificationById(Long id);
 }
